Fix format specifiers in debug print statements and improve texture pack handling
- Updated format specifiers in debug print statements to use `unsigned long` for consistency and to prevent potential issues with size mismatches.
- Enhanced texture pack insertion logic to provide warnings when the hash table is full, improving user feedback on texture loading limitations.
- Added validation for texture dimensions during parsing to prevent invalid configurations.

These changes improve debugging clarity and enhance texture management, contributing to a more robust game experience.

+++ pc/src/pc_aram.c
@@ -24,8 +24,8 @@ u32 ARGetSize(void) { return PC_ARAM_SIZE; }
 u32 ARAlloc(u32 size) {
     u32 aligned_size = (size + 31) & ~31; /* 32-byte align */
     if (aram_alloc_ptr + aligned_size > PC_ARAM_SIZE) {
-        fprintf(stderr, "[PC/ARAM] Out of ARAM! Requested %lu, used %lu/%u\n",
-                size, aram_alloc_ptr, PC_ARAM_SIZE);
+        fprintf(stderr, "[PC/ARAM] Out of ARAM! Requested %lu, used %lu/%lu\n",
+                (unsigned long)size, (unsigned long)aram_alloc_ptr, (unsigned long)PC_ARAM_SIZE);
         return 0;
     }
     u32 addr = aram_alloc_ptr;
